requirements.txt
pymongo
python-dotenv
beautifulsoup4
flask